[DEV] ✨ Refactos + Finished MongoDB connection

+++ backend/src/main/scala/api/services/UserSvc.scala
@@ -81,7 +81,7 @@ object UserSvc {
 
     // Check if existing valid token then provide a valid one
     genAccessToken  <- genString(100)
-    genExpireAt     <- Clock[IO].realTime.map(x => new Timestamp(x.toMillis + (ConfigLoader.tokenDuration.toLong * 1000)))
+    genExpireAt     <- Clock[IO].realTime.map(x => new Timestamp(x.toMillis + (ConfigLoader.tokenDuration * 1000)))
     genRefreshToken <- genString(100)
     inDBToken       <- TokenMod.select(fr"user_id = ${validatedUser.id}")
     token           <- inDBToken match {
@@ -134,7 +134,7 @@ object UserSvc {
                    case List(token) =>
                      for {
                        genAccessToken  <- genString(100)
-                       genExpireAt      = new Timestamp(nowTimestamp + (ConfigLoader.tokenDuration.toLong * 1000))
+                       genExpireAt      = new Timestamp(nowTimestamp + (ConfigLoader.tokenDuration * 1000))
                        genRefreshToken <- genString(100)
                        out             <- TokenMod.update(
                                             token
